+++ example.env
@@ -3,4 +3,5 @@ PMPERMIT_ENABLED="true"
 MONGODB_URL="mongodb://tuhin:tuhin@localhost:27017/"
 DEFAULT_TR_LANG="en"
 ENABLE_DELETE_ALERT="true"
+TIMEZONE="Asia/Kolkata"
 OCR_SPACE_API_KEY=""
